enable build of ldapscripts for trixie
Enable ldapscripts in image .inc and pkg_dirs.  Redo patch to reset pwd
in ldapadduser:

Avoiding failure with code 20 when including the field "pwdReset",
used to reset the password for the used add. This fixes an issue
in the puppet bootstrap manifest, during the task 'create ldap
admin user'.

In Trixie, the field seems to be set by default when user is
created. This sugests that this patch could be redundant, but since
I couldn't find the modification that introduced this new behaviour,
I'm including the code 20 as a value accepted and keeping the patch.

Test plan:
PASS  (f/trixie branch) Run bootstrap until the place where the
      code is executed
PASS  build bullseye and trixie on master branch
PASS  Inspect bullseye and trixie build outputs (a trixie change)

Story: 2011360
Task: 52986

Change-Id: I04170c058db09842d230fecef43a5454b90282ac
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>
Signed-off-by: Michel Thebeau <Michel.Thebeau@windriver.com>