Limit memory usage of k8s-infra-stx cgroup
This change enforces a total memory limit on all pods running in the
k8s-infra-stx cgroups. OS will start to evict pods in this cgroups when
the limit is reached through the OOM (Out-of-Memory) Killer process.

The limit value is the same as the memory reserved for the system (that
changes depending on the installation scenario).

TEST PLAN:
PASS: Build and install ISO.
PASS: AIO-SX: Verify that kubepods folder in the k8s-infra-stx cgroup
      has the memory limit set (property: memory.limit_in_bytes).
PASS: AIO-SX: Verify that the limit value is equal to the system's
      reserved memory.
PASS: AIO-SX: Verify that no OOM Killer process has been invoked.

Story: 2011626
Task: 53315

Change-Id: I53ed5548eb3888defb4100aa27073789a06d14df
Signed-off-by: Reynaldo P Gomes <reynaldo.patronegomes@windriver.com>