Fix remove dh-python and setuptool patches.
A recent commit attempted to remove the dh-python and setuptools
libraries, but it was reverted because it broke the StarlingX build.
The root cause of the failure was that the packages were removed from
the directories, but they were still listed in some files.

This commit fixes the issues introduced in the previous commit and
also adds two additional commits required to correctly remove the
libraries: one that fixes the circular dependency created after the
library removal, and another that adds the upstream setuptools to be
downloaded, as it is a dependency of other packages.

References:
 - https://review.opendev.org/c/starlingx/integ/+/968021
 - https://review.opendev.org/c/starlingx/tools/+/971248
 - https://review.opendev.org/c/starlingx/root/+/969323

Test plan:
PASS: Run ./stx-init-env --hard-reset
PASS: Successfully run downloader command
PASS: Successfully run build-pkgs command
PASS: Successfully run build-image command
PASS: AIO-SX fresh install.

Story: 2011601
Task: 53184

Change-Id: Ia2827fe8d64106caf9a7609846b38d7e160970b2
Signed-off-by: edias <edson.dias@windriver.com>